[starnix] Redirect build container package to the default toolchain
When an android_starnix_container template is instantiated under some
toolchain that is not the default, that would result in gn gen errors
as it tries to reference dependency packages under the wrong toolchain.

We already have some uses of this template but it just so happened that
none of the .gn files instantiating the template was referenced from an
alternative toolchain so it appeared to work.

Change-Id: I6280a6a2221ada89d1e04fb642ea8fc73cb5dac9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887976
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>