[starnix] Retry reading coredump report on partial nodes.
The process for creating a coredump report is roughly:

1. create an inspect node in the bounded list implementation
2. record the pid
3. record the argv

The kernel's inspect output is complete before (1) and after (3),
but any reads in the middle will see a coredump report which is missing
one or more properties.

Ideally we'd be able to add nodes to the bounded list only once they've
been fully populated but that would require API changes for that
library. In the meantime, relax the test assertions and retry
snapshotting if we have a report node but are missing a property.

Multiply: starnix_coredump_diagnostics_test
Bug: 131002
Bug: 131004
Change-Id: I1cf1541103ce57ad71b662577d45f5d53edda7c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888437
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Adam Barth <abarth@google.com>