[tee_manager] Add core shard target
Adds a target for the "base" core shard.

Before, the base core shard would be included in a board-specific
shard and extended with `use` declarations to route protocols to
v1 components or the shell realm. The derived shards only have `use`
declarations, and nothing else.

Once https://fxrev.dev/881091 is submitted, the `use` declarations
have no effect and can be removed. That makes the derived shards
equal to the base shard, so a follow up change will replace them
with a reference to this new target.

Once the derived shards are removed, the `base` core shard can
be renamed.

Bug: 104628
Change-Id: Iadcc044c6b7fdbc06a5860ee1acf08697a3eb679
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/885774
Reviewed-by: Marina Ciocea <marinaciocea@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>