[fastboot] add Rust install_from_usb()
Adds Rust implementation for install_from_usb(), which will be hooked
into a `fastboot oem` command in a follow-up CL.

The implementation itself isn't too bad, but trying to use dependency
injection to mock out some external functions got a bit messy. There are
some potential cleanups to investigate in the future, but this seemed
like a good start and the tests are passing.

Bug: b/282797139
Change-Id: I9f8e273c12f17824543f09c26aba8fd2312678db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888244
Reviewed-by: Sergii Parubochyi <sergiip@google.com>
Commit-Queue: David Pursell <dpursell@google.com>