[wlan][hw-sim] Implement realm factory configuration
Extends the realm factory FIDL protocol to allow the test suite to
configure the test realm.

Note that this was tested on a future commit together with the hw-sim
tests that actually require configuration, and the test in
`realm_factory/test.rs` was deleted.

Bug: 125740
Change-Id: I912f36be26c27fcce5d3e12376f5d384a5b0a8d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882913
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>
Commit-Queue: John Tan <johntan@google.com>