[kernel][arm64][mmu] Add some ASID asserts
When context switching, assert that the asid_ of the ArmArchVmAspace we
are switching to is in the valid range for user ASIDs. Also add some
static asserts that user ASIDs do not overlap with the UNUSED and GLOBAL
sentinels.

When the ArmArchVmAspace is destroyed, the asid_ is set to
MMU_ARM64_UNUSED_ASID (0), so this change prevents accidentally sharing
TLB entries across processes. This should not happen anyway because no
threads should be running after the ArchVmAspace of the process has been
destroyed, but the additional safety check does not hurt.

Run-All-Tests: true
Change-Id: I10bfd427d9b0498060f9ca0e688b655660f63d79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884754
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>