[fxfs] Remove unnecessary stream create
Use a VMO directly in PagedObjectHandle and FxBlob instead of wrapping
the VMO in a VmoDataBuffer. The VmoDataBuffer creates a stream object
for reading and writing to the VMO but PagedObjectHandle and FxBlob
don't use those methods. Removing the VmoDataBuffer wrapper will
remove the stream creation which will speed up opening files and blobs.

Change-Id: Ia2ffdd51d7542f35774d984bc60c41692df98539
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889898
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>