[assembly] Always generate a base_driver manifest
The Driver Index unconditionally expects a base driver manifest file to
be present in `/boot`, even if it is empty. This CL amends the assembly
builder so that a base driver manifest file is always generated.

Also clarified the TODO(https://fxbug.dev/98103) comment.

Change-Id: I117fe61d0b06d27a0c23d819ae651c8caf65771c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888239
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>