[fxblob] Create core_with_fxblob product
We're transitioning from fxblob being a board config to a product
config, now that technical roadblocks in assembly have been removed
allowing images configs to come from the product.

For now, this config is more or less a stub and won't work on its own.
It needs to be used in conjunction with the x64-fxfs board.  Eventually
it will be self-contained.  It's being defined now so we can do a soft
transition of existing CQ builders.

Note that the x64-fxfs board is unchanged by this patch and works with
or without the new product.

Bug: 291976815
Change-Id: Iaee8a6457159a19cbc28a050dc995c0a2f644ecf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890177
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>