[netstack3] FakeNetwork holds a weak device ID
In some tests we create a FakeNetwork that holds strong references to
ethernet devices created in the test. This can make it tricky to
properly clean up the devices at the end of the test and avoid dangling
strong references. Change these tests so that FakeNetwork instead holds
a weak device ID that it upgrades as needed.

Change-Id: Ia57f996a8887eef332210e2020940725ecbf1396
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889874
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>