[storage] More Controller multiplexing fixes
This is another set of generic fixes to either use the correct
FIDL channel or connect to the fuchsia.device/Controller path.

This causes a handful of storage tests to pass with multiplexing
disabled, we are down to only a couple of failing tests now.

Bug: 127870
Change-Id: I29500ee971288f169238307196bab943cc56807e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889879
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>