[starnix] Cleanup waitpid tests global state.
This CL modifies the WaitpidExitSignalTests so that the global state
they use is cleared before each test is run.

These tests spawn child processes/threads and wait on them, checking
that they issue the right signals when they exit.

The previous code had a bug in which it was possible for a signal to be
received before the fork was created, causing the global state to
contain bogus signal information.

To mitigate that, we now clean up the global state after we fork, making
it impossible (modulo starnix bugs) to receive unwanted signals.

BUG=131156

Change-Id: I6e41336162fccdd7e7ee510fa056474f3559ae5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890116
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>