[fuchsia-sync] Implement RwLock
We use two futexes, one for readers and one for writers. Once a writer
arrives, we mark a bit in our state so that readers do not starve
writers.

This RwLock does not track thread ownership for priority inheritance.
It's difficult to do that tracking in general for read/write locks, but
a future CL could improve this situation for states in which a single
thread has exclusive (writer) access to the lock.

Additionally, zx_futex_wake does not indicate whether we succeeded in
waking up a thread, which means we always need to broadcast the unlock
signal to all the readers whenever there are both readers and writers
waiting for the lock. If a writer wins the race, we will have woken up a
lot of readers without much benefit.

This lock will be used in Starnix. There is an argument that we should
not use RwLocks in Starnix. We will study our use of RwLocks in Starnix
and likely switch to mutexes in many cases. However, there are some
cases where a RwLock is needed by our current implementation approach.

Bug: 88603
Change-Id: Ib2fd45d5ead3d258e369a49fb18cfa0077f71dbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889013
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Adam Barth <abarth@google.com>