[vim3][dwmac] Fix interrupt behavior
If an ethernet cable is plugged in on boot, the DMA_STATUS_GLPII bit may
be set. That will cause the interrupt line to always be high, so with an
edge-triggered interrupt the interrupt wait may take forever.

There are two parts to this fix:
* Mark the interrupt as level-triggered. That matches the documented
  behavior of sbd_intr_o.
* Read DW_MAC_MAC_LPI_CONTROL as needed to clear DMA_STATUS_GLPII.

Fixed: 131163
Change-Id: I39bc5ec67f5c0e2c51d82f2a7cb9b20c5fcfa505
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891177
Reviewed-by: Braden Kell <bradenkell@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>