[netstack3] Rename SocketStateEntry::try_insert
Rename try_insert to try_insert_with since it takes a callback that is
used to construct the inserted value. Use the name for a simplified
version that provides a static id. Note that try_insert_with is not long
for this world, but add tests for its behavior on error (which seems to
have been previously untested).

Change-Id: Ia5ad2d2b5619b12684f117f150dc88364b8e441a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891205
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>