[netlink] Fix panic opportunity
`netlink` assumed that `AddressStateProvider.WatchAssignmentState` will
always be handled before the address is removed for reasons not related
to add-errors. This is not true because in theory, the request could be
sitting in the channel when the address is removed.

This change introduces an "on-added" event to the ASP that the server
will emit when the address is successfully added.

Fixed: b/292622684
Change-Id: Ib72a8a7610637731a518241d9d048be9bd4fdf1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890667
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>