[kernel][arch] Support values of TCR_EL1.TnSz != 24
On ARM, TCR_EL1.T{0,1}Sz are configured as
`64 - the size of the virtual address range`. These values must be
accounted for in the paging traits, as beyond restricting the size of
the virtual address space, they may further restrict the number of
paging levels (all else held equal) - and the number of paging levels is
a static determination made by the traits.

The relevance of this support is that we currently support lower ARM
virtual address spaces of size 39 and 31 (both giving three-level
paging) in physboot and the VMM.

Change-Id: Ia8358c6c1c3fda20d866e6dfda92cb87e97044d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887018
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>