[netstack3] Hold v4 & v6 socket in UDP v4 map
Modify the datagram spec impl for UDP to vary the type of receiving
identifier by the IP version of the demultiplexing map. Allow holding
only a SocketId<Ipv6> in the IPv6 demultiplexing map, but an
Either<SocketId<Ipv4>, SocketId<Ipv6>> (spelled as EitherIpSocket<Udp>)
in the IPv4 demultiplexing map.

This is a key first step towards enabling cross-IP-version reception of
traffic since it will enable IPv6 receivers to be found for IPv4
packets. While this change updates the types to make it possible, there
is no path for adding IPv6 sockets to the IPv4 map so all socket IDs are
still assumed to be IPv4.

Bug: 21198
Change-Id: I8125049814bdf9aeccd34c30a890a5e94f252250
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891173
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>