[symbolizer] Support backtrace prettifying
Add a new flag to the symbolizer, --prettify-backtrace. When enabled,
backtraces in {{{reset:begin}}} and {{{reset:end}}} pairs will be output
in a compressed and prettified format, allowing more readability.

The flag is not enabled by default and there should be no behavior
change when the flag is not enabled.

Bug: 59736
Change-Id: I2308344f828c97f187ee88437ce86efc88e9e5bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892726
Commit-Queue: Dangyi Liu <dangyi@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>