[starnix] Start up synthetic UTC clock if real one is not started
Starnix relies on a running UTC time source as many Linux programs
expect CLOCK_REALTIME to advance in order to make progress. Previously
we waited for the UTC clock handle provided to the Starnix kernel to
start running before we started any Linux programs. This causes issues
as on some devices the Fuchsia timekeeper service is not able to find a
reliable time source - such as emulators without internet access - and
in some configurations we may want to run Linux programs to run
configuration required to find a clock source. This modifies the startup
procedure in Starnix to first check if the UTC clock is running and if
it isn't use a synthetic "clock" that starts at the UTC backstop time
and advances at the same rate as the monotonic clock. Once the real UTC
clock starts running (if it ever does), we replace the synthetic time
source with the real UTC clock.

Bug: 131200

Change-Id: Icbad1919b4e11f8ae54762c2435c6b36f1ebdc1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891825
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>