[ld] Load main executable
This makes the dynamic linker actually handle loading the
executable.  This includes setting some of the groundwork for
loading modules in general, but does not yet do any relocation.
After loading, it hands off to the executable with the same
initial conditions it started with.

In the POSIX incarnation, "loading" is actually just working with
the executable image already loaded by the system program loader
and found via auxv entries.  The test harness is enhanced to
mimic the system program loader more fully by loading the
executable and filling out the related auxv entries.

The build and test machinery is prepared for a wider set of test
modules to be built in the future.  This is a step closer to
generalizing the same tests to be done in the current in-process
fashion as well as in multiple out-of-process fashions.

Bug: 130483
Change-Id: Ibed96845fa1eb7ca20e6ca7f6498b164df929de3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888711
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Alex Brachet <abrachet@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>