[netstack] Warn for not-found thread profile
Currently we log an error whenever we fail to set the thread profile
for the threads the netstack runs on, but some failures are expected
when no matching thread profile is configured with the requested
profile name. Downgrades that specific not-found case to a warning
instead of an error to mitigate logspam.

Bug: b/294198743
Change-Id: I87053c24f1b9c2a6564043cdc380290d07e17c6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893960
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>