[starnix][pidfd] Integrate with waitid
Previously, pidfd was holding onto the Task with a WeakRef<Task>, but
that doesn't work for waitid, which uses the pidfd to specify the task
after the task has exited. After this CL, we use the pid itself, which
is sufficient for now but will need to be revisited in the future. See
the comment on `PidFdFileObject::pid` for more details.

Test: pidfd_open04
Bug: 119476
Fixed: b/287120994
Change-Id: I370173a8a034d8346b614d4e6999d66a7c9048b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893766
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Vickie Cheng <vickiecheng@google.com>