[fxfs] make XtsCipherSet on demand
Instead of pre-making the XtsCipherSet, make a new one every time we
need the keys.

Premaking the set means we create the Xts128 struct containing two of
the Aes256 structs, which each take up about 1k of memory regardless of
if we have AES intrinsics on the board we are running on. This is an
enormous amount of memory per node, so instead we pay a cpu cost every
time we encrypt. The cost is very small when intrinsics are enabled
(just a copy or two), thought it is slightly larger when they aren't.

Change-Id: If91a3c7a1de5853b43253ab326e8f0f7f40ee530
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893460
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>