[owners] Extend toplevel OWNERS rule.
I've confirmed that `per file OWNERS=...` at the top level only
grants ownership over the single OWNERS file at the root of the
source tree. But we'll want to mess around with OWNERS files deeper
in the tree as well, so I'm extending that.

Change-Id: Ic6bed67eed66bd91434f4a3fa24db84658188047
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894739
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>